requirements.txt
beautifulsoup4==4.12.3
bson==0.5.10
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.2
opencv-python==4.11.0.86
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
soupsieve==2.6
urllib3==2.3.0
